refactor: extract card number and expiry date input logic into custom hooks
