+++ README.md
@@ -81,5 +81,5 @@
   - [x] react-router 사용
 - Custom Hooks
   - [ ] 사용하여 Form 관리 로직을 UI와 분리하기 (구현 완료 후)
-    - [ ] 입력값, 에러 상태, onChange, onBlur 패턴 공통 커스텀훅으로 추출 (검증 메서드 주입)
+    - [x] 입력값, 에러 상태, onChange, onBlur 패턴 공통 커스텀훅으로 추출 (검증 메서드 주입)
     - [ ] focus 관련 훅들 커스텀 훅으로 모아놓기
